Refactor newUser a bit
