+++ src/components/organisms/Setup.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Field } from "redux-form";
 import { List, Button, Card } from "semantic-ui-react";
 
-import renderInput from "./utils/renderInput";
+import renderInput from "../utils/renderInput";
 
 // Explicitely decompose props here.
 const Setup = ({ loading, handleSubmit }) => (
